Add files via upload
These are two php files for storing images in the database and pulling images to display to the screen. I also have another php file to actually pull the images from the database. Once I have the information to connect to the MySQL server I will integrate these with the upload button on the html page